requirements.txt
librosa
numpy
matplotlib
pandas
scipy
torch
black
ffmpeg-python
pydub
